Remove HelpSearchViewActivity
Resolved issue with Help activity not changing to DarkTheme. Need to fix back button.
