+++ app/src/main/java/com/madonasyombua/growwithgoogleteamproject/actvities/HelpActivity.java
@@ -3,18 +3,32 @@ package com.madonasyombua.growwithgoogleteamproject.actvities;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.net.Uri;
-import android.preference.PreferenceManager;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.preference.PreferenceManager;
+import android.speech.RecognizerIntent;
+import android.support.design.widget.Snackbar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.text.TextUtils;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.widget.ImageButton;
 import android.widget.TextView;
+
 import com.madonasyombua.growwithgoogleteamproject.R;
 import com.madonasyombua.growwithgoogleteamproject.ui.SharedPref;
+import com.miguelcatalan.materialsearchview.MaterialSearchView;
+
+import java.util.ArrayList;
+
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
 public class HelpActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener {
+
+    // SearchView
+    private MaterialSearchView searchView;
 
     @BindView(R.id.backToMain)
     ImageButton backToMain;
@@ -29,6 +43,7 @@ public class HelpActivity extends AppCompatActivity implements SharedPreferences
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+
         //Theme can only be changed before setContentView is called.
         //Therefore, I am changing the theme on here.
         sharedPref = new SharedPref(this);
@@ -40,9 +55,42 @@ public class HelpActivity extends AppCompatActivity implements SharedPreferences
         prev_State = sharedPref.loadNightModeState();
 
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_help);
+        setContentView(R.layout.activity_help_searchview);
         ButterKnife.bind(this);
         setCorrectTheme();
+
+        // Toolbar with Search Icon
+        Toolbar toolbar = (Toolbar) findViewById(R.id.tb_help);
+        setSupportActionBar(toolbar);
+
+        searchView = (MaterialSearchView) findViewById(R.id.search_view);
+        searchView.setVoiceSearch(true);
+        searchView.setCursorDrawable(R.drawable.color_cursor_white);
+        searchView.setSuggestions(getResources().getStringArray(R.array.search_suggestions));
+        searchView.setOnQueryTextListener(new MaterialSearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                Snackbar.make(findViewById(R.id.container), "Query: " + query, Snackbar.LENGTH_LONG).show();
+                return false;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String newText) {
+                return false;
+            }
+        });
+
+        searchView.setOnSearchViewListener(new MaterialSearchView.SearchViewListener() {
+            @Override
+            public void onSearchViewShown() {
+                // Do something something
+            }
+
+            @Override
+            public void onSearchViewClosed() {
+                // Do something something
+            }
+        });
 
         backToMain.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -53,7 +101,6 @@ public class HelpActivity extends AppCompatActivity implements SharedPreferences
             }
         });
 
-        //for now we will use this
         faq.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -87,6 +134,7 @@ public class HelpActivity extends AppCompatActivity implements SharedPreferences
             }
         });
     }
+
     @Override
     protected void onDestroy() {
         super.onDestroy();
@@ -110,6 +158,38 @@ public class HelpActivity extends AppCompatActivity implements SharedPreferences
             startActivity(new Intent(this, this.getClass()));
             finish();
         }
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.help_menu, menu);
+        MenuItem item = menu.findItem(R.id.action_search);
+        searchView.setMenuItem(item);
+        return true;
+    }
+
+    @Override
+    public void onBackPressed() {
+        if (searchView.isSearchOpen()) {
+            searchView.closeSearch();
+        } else {
+            super.onBackPressed();
+        }
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        if (requestCode == MaterialSearchView.REQUEST_VOICE && resultCode == RESULT_OK) {
+            ArrayList<String> matches = data.getStringArrayListExtra(RecognizerIntent.EXTRA_RESULTS);
+            if (matches != null && matches.size() > 0) {
+                String searchWord = matches.get(0);
+                if (!TextUtils.isEmpty(searchWord)) {
+                    searchView.setQuery(searchWord, false);
+                }
+            }
+            return;
+        }
+        super.onActivityResult(requestCode, resultCode, data);
     }
 
     @Override
